Apply modified Debian 05-gniibe-fix-13.patch, to fix heap-based buffer overflow in the do_msg() function
https://sourceforge.net/p/mcpp/bugs/13/ and CVE-2019-14274

Description: Fix for a bug reported to sourceforge.net #13
   by fixing error messages.
   Also, fix erroneous messages.
Author: NIIBE Yutaka

The buffer overflow was caused by do_msg() treating illegal control characters
in the error message output as tokenized macros that needed to be expanded for
printing.

This fixes the problem by just not printing the preprocessed source code which
contains those illegal control characters.

The patch has been reworked to change the error messages to better match those
that are output when unterminated strings are encountered. It also changes the
other patched messages to restore the 'l' in "%02lx" because do_msg() passes a
long integer to mcpp_fprintf().