Fix heap-use-after-free in substitute() causes segfault
https://sourceforge.net/p/mcpp/bugs/14/

This is actually nothing to do with a heap-use-after-free!

It is caused by get_repl() corrupting the invocation of the 9th or 32nd macro
parameter when it is used at the end of the replacement text, because it sees
the encoded 0x09 or 0x20 formal parameter index as a trailing space to delete.

That corruption then goes on to cause the undefined behavior which results in
a segfault.